Don't create settings file until serialization
Otherwise you could end up with an empty file and lose all your settings if the JsonTextWriter fails. And good practice to avoid race conditions with the watcher especially.

Uses more memory but the file is so small it's not a reasonable concern.

Potential fix for https://github.com/danielchalmers/DesktopClock/issues/7#issuecomment-1830772175